temp

+++ src/main/java/shop/bookbom/shop/domain/order/dto/response/BeforeOrderBookResponse.java
@@ -9,5 +9,5 @@ public class BeforeOrderBookResponse {
     private String imageUrl;
     private String title;
     private Integer quantity;
-    private Long cost;
+    private Integer cost;
 }
